pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hybrid-pdf-ocr-macgui"
version = "0.1.0"
description = "Hybrid PDF OCR system with Document AI, Donut, TrOCR, and TATR for macOS"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["ocr", "pdf", "document-ai", "donut", "trocr", "macos", "gui"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS :: MacOS X",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Core
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "click>=8.1.0",

    # GUI
    "PyQt5>=5.15.0",
    "PyQt5-Qt5>=5.15.0",

    # Google Cloud
    "google-cloud-documentai>=2.20.0",
    "google-cloud-storage>=2.10.0",
    "google-auth>=2.23.0",

    # ML Framework
    "torch>=2.1.0",
    "transformers>=4.35.0",
    "onnxruntime>=1.16.0",
    "timm>=0.9.0",
    "peft>=0.7.0",  # For LoRA

    # PDF/Image Processing
    "PyMuPDF>=1.23.0",  # fitz
    "pdf2image>=1.16.0",
    "Pillow>=10.0.0",
    "pypdfium2>=4.24.0",

    # Computer Vision
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "scikit-image>=0.22.0",

    # OCR & Layout
    "pytesseract>=0.3.10",

    # Document Export
    "python-docx>=1.1.0",
    "markdown>=3.5.0",

    # Evaluation
    "editdistance>=0.6.2",
    "jiwer>=3.0.0",

    # Utilities
    "tqdm>=4.66.0",
    "rich>=13.7.0",
    "tenacity>=8.2.0",
    "aiofiles>=23.2.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.10.0",
    "ruff>=0.1.0",
    "mypy>=1.6.0",
    "pre-commit>=3.5.0",
]

table = [
    "table-transformer>=1.0.0",
    "detectron2>=0.6",
]

math = [
    "pix2tex>=0.1.2",
    "sympy>=1.12",
]

all = [
    "hybrid-pdf-ocr-macgui[dev,table,math]",
]

[project.scripts]
hybrid-ocr = "gui.desktop_app:main"

[project.urls]
Homepage = "https://github.com/yourusername/hybrid-pdf-ocr-macgui"
Repository = "https://github.com/yourusername/hybrid-pdf-ocr-macgui"
Issues = "https://github.com/yourusername/hybrid-pdf-ocr-macgui/issues"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
namespaces = false

[tool.black]
line-length = 100
target-version = ["py312"]
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py312"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=term-missing"
